refactor: changed expressionParser to expression_parser
